api: rename doip_message to doip_payload, remove version bytes transitions
