+++ src/doip_payload/activation_code.rs
