+++ src/doip_payload/diagnostic_message.rs
