+++ src/doip_payload/diagnostic_message_nack.rs
