+++ src/doip_payload/diagnostic_nack.rs
