+++ src/doip_payload/power_information_response.rs
