+++ src/doip_payload/power_mode.rs
