+++ src/doip_payload/vehicle_announcement_message.rs
