+++ src/doip_payload/vehicle_identification_request_eid.rs
