+++ src/doip_payload/vehicle_identification_request_vin.rs
